requirements.txt
absl-py==1.0.0
appdirs==1.4.4
backcall==0.1.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.9.0
bs4==0.0.1
CacheControl==0.12.10
cachetools==4.2.4
cachy==0.3.0
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.7
cityscapesScripts==2.2.0
cleo==0.8.1
clikit==0.6.2
coloredlogs==15.0
colorlog==4.1.0
crashtest==0.3.1
cryptography==35.0.0
cycler==0.10.0
Cython==0.29.22
dataclasses==0.8
decorator==4.4.2
distlib==0.3.3
easydict==1.9
filelock==3.4.0
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
greenlet==0.4.15
grpcio==1.42.0
html5lib==1.1
humanfriendly==9.1
idna==3.3
importlib-metadata==1.7.0
importlib-resources==5.4.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.0
jeepney==0.7.1
keyring==21.8.0
kiwisolver==1.3.1
lockfile==0.12.2
Markdown==3.3.5
matplotlib==3.3.4
msgpack==1.0.0
ninja==1.10.0.post2
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
packaging==20.9
pandas==1.0.3
parso==0.7.0
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.8.1
platformdirs==2.4.0
poetry==1.1.11
poetry-core==1.0.7
prompt-toolkit==3.0.16
protobuf==3.19.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.21
Pygments==2.6.1
pylev==1.4.0
pynvim==0.4.1
pyparsing==2.4.7
pyquaternion==0.9.9
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
scipy==1.5.4
seaborn==0.10.1
SecretStorage==3.3.1
Shapely==1.8.0
shellingham==1.4.0
six==1.16.0
soupsieve==2.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tomlkit==0.7.2
torch==1.2.0
torchvision==0.4.0
tqdm==4.59.0
traitlets==4.3.3
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.7
virtualenv==20.10.0
vmem==1.0.1
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
yacs==0.1.8
zipp==3.6.0
